services/chat-processor/requirements.txt
kafka-python
boto3
psycopg2-binary
requests==2.31.0
cryptography==41.0.7
